requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiohttp==3.9.5
aiofiles==23.2.1
pytz==2024.1
python-dateutil==2.8.2
typing-extensions==4.12.2